[graphite] Clean-up Uniform header, pack fields and add type asserts
Bug: b/317232480
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/792776
Commit-Queue: Michael Ludwig <michaelludwig@google.com>

+++ src/gpu/graphite/UniformManager.cpp
@@ -549,7 +549,7 @@ void UniformManager::checkExpected(SkSLType type, unsigned int count) {
 
     SkASSERT(fExpectedUniforms[fExpectedUniformIndex].type() == type);
     SkASSERT((fExpectedUniforms[fExpectedUniformIndex].count() == 0 && count == 1) ||
-             fExpectedUniforms[fExpectedUniformIndex].count() == count);
+             fExpectedUniforms[fExpectedUniformIndex].count() == (int) count);
     SkDEBUGCODE(fExpectedUniformIndex++;)
 }
